+++ R/phenotype.R
@@ -14,7 +14,7 @@
 #'
 #'
 #' @export
-#' @import R6
+#' @importFrom R6 R6Class
 trait <- R6::R6Class(
   "trait",
   lock_objects = FALSE,
@@ -168,7 +168,7 @@ trait <- R6::R6Class(
 #' phenotyper object is used to phenotype populations
 #'
 #' @export
-#' @import R6
+#' @importFrom R6 R6Class
 phenotyper <- R6::R6Class(
   "phenotyper",
   lock_objects = FALSE,
@@ -384,8 +384,8 @@ phenotyper <- R6::R6Class(
     #' results
     #' @details
     #' phenotypic values for individual i repetition j is calculated as follow:
-    #' \deqn{y_{ij} = \mu + g_i + e_ij + offset}
-    #' \deqn{e_ij \sim N(0, \sigma_e^2)}
+    #' \deqn{y_{ij} = \mu + g_i + e_{ij} + offset}
+    #' \deqn{e_{ij} \sim N(0, \sigma_e^2)}
     #' where g_i is the genetic value of the individual i
     #' @examples
     #'
@@ -466,7 +466,7 @@ phenotyper <- R6::R6Class(
       list(
         data = dta,
         cost = cost
-      )
+)