+++ NEWS.md
@@ -1,5 +1,10 @@
 # breedSimulatR 0.3.1
 
+
+## fix
+
+- fix wrong documentation in `phenotype` and `haplotype` classes.
+
 ## Performence optimization
 
 - `population$addInd` methods will now only check that the specie's names match between the population and the individual instead of checking if the specie objects are exactly the same. This can introduce a bug if two differents species with the same name are use in a simulation however, this cases will be very unlikely and the performance gain is quite high.